Removed declarations of direct dependencies of <type>pom</type>.
This was an attempt to control reactor build order.  This was not the correct
solution to our problem.  We're not exactly sure which step solved our build
problems, but we have recently resolved several other issues, including:

1) Fixed some incorrect parent <relativePath> declarations
2) Resolved some complicated cyclic dependencies
3) Removed the use of some properties in the versions used for <scope>import</scope>

We are no longer able to reproduce the original issue that prompted this
attempted solution.  Specifically, we are now able to build our entire set of
943 projects in a single self-contained reactor build while in offline mode.